NetCoMi: Network Comparison for Microbial Compositional Data

NetCoMi offers functions for constructing, analyzing, and comparing microbial association networks as well as dissimilarity-based networks for compositional data. It also includes a function for constructing differential association networks.

The main functions are netConstruct, netAnalyze, netCompare, and diffnet

See also

Author

Maintainer: Stefanie Peschel stefanie.peschel@mail.de

Acknowledgments:

Anne-Laure Boulesteix, Christian L. Müller, and Martin Depner (inspiration and theoretical background)

Anastasiia Holovchak (package testing and editing)