Function for Generating Vector Names

The function generates names for a vector that contains elements of the lower triangle of a matrix. The R code has been copied from getNames (not exported function of the discordant package) and changed such that the names are generated by columns (and not by rows as in the original function, which produces implausible results).

Usage

.getVecNames(x)

Arguments

x

symmetric matrix for which the column and row names should be returned as a vector