Multiple testing adjustment

The functions adjusts a vector of p-values for multiple testing

Usage

multAdjust(
  pvals,
  adjust = "adaptBH",
  trueNullMethod = "convest",
  pTrueNull = NULL,
  verbose = FALSE
)

Arguments

pvals

numeric vector with p-values

adjust

character specifying the method used for adjustment. Can be "lfdr", "adaptBH", or one of the methods provided by p.adjust.

trueNullMethod

character indicating the method used for estimating the proportion of true null hypotheses from a vector of p-values. Used for the adaptive Benjamini-Hochberg method for multiple testing adjustment (chosen by adjust = "adaptBH"). Accepts the provided options of the method argument of propTrueNull: "convest"(default), "lfdr", "mean", and "hist". Can alternatively be "farco" for the "iterative plug-in method" proposed by Farcomeni (2007).

pTrueNull

proportion of true null hypothesis used for the adaptBH method. If NULL, the proportion is computed using the method defined via trueNullMethod.

verbose

if TRUE, progress messages are returned.

References

farcomeni2007someNetCoMi